Minor code restructuring and cleanup.

+++ Tonuino_Main/src/shared/Tonuino_Potentiometer.cpp
@@ -1,7 +1,7 @@
 
 #include "Tonuino_Potentiometer.h"
 
-void TonuinoPotentiometer::setup(int potiPin, int minScale, int maxScale)
+void TonuinoPotentiometer::setup(uint8_t potiPin, uint8_t minScale, uint8_t maxScale)
 {
 	pin = potiPin;
 	min = minScale * 2;
@@ -12,8 +12,8 @@ void TonuinoPotentiometer::setup(int potiPin, int minScale, int maxScale)
 
 bool TonuinoPotentiometer::read()
 {
-    int valueRaw = analogRead(pin);
-    int value = map(valueRaw, 0, 1024, min, max);
+    uint16_t valueRaw = analogRead(pin);
+    uint8_t value = map(valueRaw, 0, 1024, min, max);
 	
     if (valueRaw < 1023 && // ignore highest raw value, as it might occur occasionally
 	   (appliedValue == 0 // for first call
