+++ Tonuino_Main/src/shared/Tonuino_Potentiometer.h
@@ -4,13 +4,13 @@
 class TonuinoPotentiometer
 {
 	public:
-	void setup(int pin, int min, int max);
+	void setup(uint8_t pin, uint8_t min, uint8_t max);
 	bool read();
-	int appliedValue = 0; 
+	uint8_t appliedValue = 0; 
 	
 	private:
-	int pin = 0;
-	int min = 0;
-	int max = 0;
-	int hysterese = 1;
+	uint8_t pin = 0;
+	uint8_t min = 0;
+	uint8_t max = 0;
+	uint8_t hysterese = 1;
 };
